backend/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
langchain-text-splitters
pydantic
python-multipart
